Split wheel encoder into header and source files
